polish: set plan as variable and remove useless service
